grpc-service/requirements.txt
anyio==3.7.1
Faker==19.6.2
grpcio==1.58.0
grpcio-tools==1.58.0
protobuf==4.24.3
